chore(dependencies): fin du refacto sur l'injection de dependence coté server
